+++ src/main/java/lotto/service/LotteryCheckService.java
@@ -1,6 +1,11 @@
-package lotto.model;
+package lotto.service;
 
 import lotto.config.WinningCondition;
+import lotto.model.BonusNumber;
+import lotto.model.Lotto;
+import lotto.model.LottoRepository;
+import lotto.model.WinningNumber;
+import lotto.model.WinningRepository;
 
 public class LotteryCheckService {
