+++ src/test/java/lotto/service/LotteryCheckServiceTest.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.service;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -7,6 +7,11 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 
 import lotto.config.WinningCondition;
+import lotto.model.BonusNumber;
+import lotto.model.Lotto;
+import lotto.model.LottoRepository;
+import lotto.model.WinningNumber;
+import lotto.model.WinningRepository;
 
 public class LotteryCheckServiceTest {
