+++ src/test/java/lotto/service/LottoMakeServiceTest.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.service;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -6,6 +6,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Test;
 
 import lotto.config.LottoRule;
+import lotto.model.Lotto;
+import lotto.model.LottoRepository;
 
 import org.junit.jupiter.api.DisplayName;
